The Chebyshev smoother inside the multigrid algorithm runs faster (around 10-15%) if we renumber the unknowns with the new `DoFRenumbering::matrix_free_data_locality` algorithm from #13628.