@kronbichler 2022-05-05
Just a small detail on the achieved improvement: On a 72-core machine the run time of the "solve" stage in the heavy case improves from 1.47 seconds to 1.26 seconds, and the time for 50 matrix-vector products with float numbers from 0.34 seconds to 0.30 seconds. Of course, the setup time gets higher, the `setup_dofs` category is now 0.96 seconds against 0.43 seconds before. (But those techniques are meant to speed up things when done very often.)